# services/kg-query-api/src/main/kotlin/kvasir/services/api/kg/query/QueryApi.kt
package kvasir.services.api.kg.query

import com.fasterxml.jackson.annotation.JsonProperty
import io.smallrye.mutiny.Uni
import jakarta.ws.rs.*
import jakarta.ws.rs.core.MediaType
import jakarta.ws.rs.core.UriInfo
import kvasir.definitions.kg.KnowledgeGraph
import kvasir.definitions.kg.Pod
import kvasir.definitions.kg.PodStore
import kvasir.definitions.kg.QueryRequest
import kvasir.definitions.kg.QueryResult
import kvasir.definitions.openapi.ApiDocConstants
import kvasir.definitions.openapi.ApiDocTags
import kvasir.definitions.rdf.JSON_LD_MEDIA_TYPE
import org.eclipse.microprofile.openapi.annotations.Operation
import org.eclipse.microprofile.openapi.annotations.media.Content
import org.eclipse.microprofile.openapi.annotations.media.Schema
import org.eclipse.microprofile.openapi.annotations.parameters.Parameter
import org.eclipse.microprofile.openapi.annotations.responses.APIResponse
import org.eclipse.microprofile.openapi.annotations.tags.Tag
import java.time.Instant
import java.util.Optional
import kotlin.jvm.optionals.getOrNull

const val QUERY_API_PATH = "/query"

@Tag(name = ApiDocTags.KNOWLEDGE_GRAPH_API)
@Path("")
class QueryApi(
    private val knowledgeGraph: KnowledgeGraph,
    private val podStore: PodStore,
    private val uriInfo: UriInfo
) {

    @Path("{podId}$QUERY_API_PATH")
    @POST
    @Produces(MediaType.APPLICATION_JSON)
    @Operation(
        summary = "Retrieve data from the KG.",
        description = "Query the knowledge graph of the specified pod using GraphQL."
    )
    fun query(
        @PathParam("podId") podId: String,
        input: QueryInputWithContext,
        @QueryParam("atTimestamp") @Parameter(
            description = "Query the state of the KG at the specified point in time.",
            required = false
        ) atTimestamp: Optional<Instant>,
        @QueryParam("atChangeRequestId") @Parameter(
            description = "Query the state of the KG when the specified change request was applied.",
            required = false
        ) atChangeRequestId: Optional<String>
    ): Uni<QueryResult> {
        val podId = uriInfo.absolutePath.toString().substringBefore(QUERY_API_PATH)
        return podStore.getById(podId).onItem().ifNull().failWith(NotFoundException("Pod not found: $podId"))
            .onItem().ifNotNull().transformToUni { pod ->
                val req = parseInput(pod!!, input, atTimestamp.getOrNull(), atChangeRequestId.getOrNull())
                knowledgeGraph.query(req)
            }
    }

    @Path("{podId}$QUERY_API_PATH")
    @POST
    @Produces(JSON_LD_MEDIA_TYPE)
    @APIResponse(
        responseCode = "200",
        description = "The query result in JSON-LD format.",
        content = [Content(example = ApiDocConstants.JSON_LD_RESPONSE_EXAMPLE)]
    )
    fun queryJsonLD(
        @PathParam("podId") podId: String, input: QueryInputWithContext,
        @QueryParam("atTimestamp") @Parameter(
            description = "Query the state of the KG at the specified point in time.",
            required = false
        ) atTimestamp: Optional<Instant>,
        @QueryParam("atChangeRequestId") @Parameter(
            description = "Query the state of the KG when the specified change request was applied.",
            required = false
        ) atChangeRequestId: Optional<String>
    ): Uni<Map<String, Any>> {
        val podId = uriInfo.absolutePath.toString().substringBefore(QUERY_API_PATH)
        return podStore.getById(podId).onItem().ifNull().failWith(NotFoundException("Pod not found: $podId"))
            .onItem().ifNotNull().transformToUni { pod ->
                val req = parseInput(pod!!, input, atTimestamp.getOrNull(), atChangeRequestId.getOrNull())
                knowledgeGraph.query(req).map {
                    it.toJsonLD(req.context)
                }
            }
    }

    private fun parseInput(
        pod: Pod,
        input: QueryInputWithContext,
        atTimestamp: Instant? = null,
        atChangeRequestId: String? = null
    ): QueryRequest {
        return QueryRequest(
            pod.getDefaultContext(),
            pod.id,
            input.query,
            input.variables,
            input.operationName,
            input.targetGraphs,
            atTimestamp = atTimestamp,
            atChangeRequestId = atChangeRequestId
        )
    }
}

interface QueryInput {


    @get:Schema(
        description = "The GraphQL query string to be executed.",
        example = "{ id ex_givenName(_: \"Bob\") ex_friends { id ex_givenName } }"
    )
    val query: String

    @get:Schema(
        description = "The name of the operation to be executed (optional, only required if the GraphQL query expresses more than one operation)."
    )
    val operationName: String?

    @get:Schema(
        description = "The variables to be used in the query."
    )
    val variables: Map<String, Any>?

    @get:Schema(
        description = "The named graphs to be targeted by the query. If no graphs are specified, all graphs are targeted."
    )
    val targetGraphs: Set<String>
}

data class QueryInputImpl(
    override val query: String,
    override val operationName: String? = null,
    override val variables: Map<String, Any>? = null,
    override val targetGraphs: Set<String> = emptySet()
) : QueryInput

data class QueryInputWithContext(
    override val query: String,
    override val operationName: String? = null,
    override val variables: Map<String, Any>? = null,
    override val targetGraphs: Set<String> = emptySet(),
    @get:JsonProperty("@context")
    @get:Schema(
        name = "@context",
        description = "The JSON-LD context for the query.",
        example = ApiDocConstants.JSON_LD_CONTEXT_EXAMPLE_2
    )
    val providedContext: Map<String, Any>? = null
) : QueryInput

internal fun throw404IfPodNotFound(podStore: PodStore, podId: String): Uni<Void> {
    return podStore.getById(podId).onItem().ifNull().failWith(NotFoundException("Pod not found: $podId")).onItem()
        .ifNotNull().transformToUni { Uni.createFrom().voidItem() }
}