# services/monolith/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Kvasir API
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value
...
paths:
...
  /{podId}/query:
    post:
      tags:
      - Knowledge Graph API
      summary: Retrieve data from the KG.
      description: Query the knowledge graph of the specified pod using GraphQL.
      parameters:
      - name: podId
        in: path
        required: true
        schema:
          type: string
      - name: atChangeRequestId
        in: query
        description: Query the state of the KG when the specified change request was
          applied.
        required: false
        schema:
          type: string
      - name: atTimestamp
        in: query
        description: Query the state of the KG at the specified point in time.
        required: false
        schema:
          $ref: "#/components/schemas/Instant"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryInputWithContext"
      responses:
        "200":
          description: OK
          content:
            application/ld+json:
              example:
                '@context':
                  ex: http://example.org/
                '@graph':
                - '@id': ex:123
                  ex:givenName: Bob
                  ex:friends:
                  - '@id': ex:456
                    ex:givenName: Alice
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResult"
  /{podId}/slices:
    get:
      tags:
      - Knowledge Graph API
      summary: List slices of the specified pod.
      description: List slices of the specified pod's Knowledge Graph.
      parameters:
      - name: podId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/ld+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SliceSummary"
    post:
      tags:
      - Knowledge Graph API
      summary: Define a new slice of the KG.
      description: "Define a new slice (subset) of the specified pod's Knowledge Graph,\
        \ based on a GraphQL-LD schema."
      parameters:
      - name: podId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/ld+json:
            schema:
              $ref: "#/components/schemas/SliceInput"
      responses:
        "200":
          description: OK
          content:
            application/json: {}
  /{podId}/slices/{sliceId}:
    get:
      tags:
      - Knowledge Graph API
      summary: Retrieve a specific slice definition..
      description: Retrieve a specific slice definition details.
      parameters:
      - name: podId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/ld+json:
              schema:
                $ref: "#/components/schemas/Slice"
    delete:
      tags:
      - Knowledge Graph API
      summary: Delete a specific slice.
      description: Delete a specific slice of the specified pod's Knowledge Graph.
      parameters:
      - name: podId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json: {}
...
  /{podId}/slices/{sliceId}/query:
    post:
      tags:
      - Knowledge Graph API
      summary: Retrieve data from a specific subset of the KG.
      description: Query a predefined slice of the specified pod's Knowledge Graph
        using GraphQL.
      parameters:
      - name: podId
        in: path
        required: true
        schema:
          type: string
      - name: sliceId
        in: path
        description: "Identifier of the Knowledge Graph slice, representing a subset\
          \ of the specified pod's Knowledge Graph."
        required: true
        schema:
          type: string
      - name: atChangeRequestId
        in: query
        description: Query the state of the KG when the specified change request was
          applied.
        required: false
        schema:
          type: string
      - name: atTimestamp
        in: query
        description: Query the state of the KG at the specified point in time.
        required: false
        schema:
          $ref: "#/components/schemas/Instant"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryInputImpl"
      responses:
        "200":
          description: OK
          content:
            application/ld+json:
              example:
                '@context':
                  ex: http://example.org/
                '@graph':
                - '@id': ex:123
                  ex:givenName: Bob
                  ex:friends:
                  - '@id': ex:456
                    ex:givenName: Alice
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResult"
  /{podId}/slices/{sliceId}/shacl:
    get:
      tags:
      - Knowledge Graph API
      parameters:
      - name: podId
        in: path
        required: true
        schema:
          type: string
      - name: sliceId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            text/turtle:
              schema:
                type: string
...
components:
  schemas:
...
    Instant:
      format: date-time
      type: string
      example: 2022-03-10T16:15:50Z
...
    QueryInputImpl:
      required:
      - query
      - targetGraphs
      type: object
      properties:
        query:
          description: The GraphQL query string to be executed.
          type: string
          example: "{ id ex_givenName(_: \"Bob\") ex_friends { id ex_givenName } }"
        operationName:
          description: "The name of the operation to be executed (optional, only required\
            \ if the GraphQL query expresses more than one operation)."
          type: string
          nullable: true
        variables:
          description: The variables to be used in the query.
          type: object
          additionalProperties: {}
          nullable: true
        targetGraphs:
          description: "The named graphs to be targeted by the query. If no graphs\
            \ are specified, all graphs are targeted."
          uniqueItems: true
          type: array
          items:
            type: string
    QueryInputWithContext:
      required:
      - query
      - targetGraphs
      type: object
      properties:
        query:
          description: The GraphQL query string to be executed.
          type: string
          example: "{ id ex_givenName(_: \"Bob\") ex_friends { id ex_givenName } }"
        operationName:
          description: "The name of the operation to be executed (optional, only required\
            \ if the GraphQL query expresses more than one operation)."
          type: string
          nullable: true
        variables:
          description: The variables to be used in the query.
          type: object
          additionalProperties: {}
          nullable: true
        targetGraphs:
          description: "The named graphs to be targeted by the query. If no graphs\
            \ are specified, all graphs are targeted."
          uniqueItems: true
          type: array
          items:
            type: string
        '@context':
          description: The JSON-LD context for the query.
          type: object
          additionalProperties: {}
          example:
            ex: http://example.org/
          nullable: true
    QueryResult:
      type: object
      properties:
        data:
          type: object
          additionalProperties: {}
          nullable: true
        errors:
          type: array
          items:
            type: object
            additionalProperties: {}
          nullable: true
...
    Slice:
      required:
      - id
      - context
      - podId
      - name
      - description
      - schema
      - shacl
      - targetGraphs
      type: object
      properties:
        id:
          type: string
        context:
          type: object
          additionalProperties: {}
        podId:
          type: string
        name:
          type: string
        description:
          type: string
        schema:
          type: string
        shacl:
          type: string
        targetGraphs:
          uniqueItems: true
          type: array
          items:
            type: string
    SliceInput:
      required:
      - context
      - name
      - schema
      - description
      - targetGraphs
      type: object
      properties:
        context:
          type: object
          additionalProperties: {}
        name:
          type: string
        schema:
          type: string
        description:
          type: string
        targetGraphs:
          uniqueItems: true
          type: array
          items:
            type: string
    SliceSummary:
      required:
      - id
      - name
      - description
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
...